# Optimize writing optional columns (part 5)