--- a/writer_go18_test.go
+++ b/writer_go18_test.go
@@ -28,6 +28,7 @@ func BenchmarkGenericWriter(b *testing.B) {
 	benchmarkGenericWriter[contact](b)
 	benchmarkGenericWriter[paddedBooleanColumn](b)
 	benchmarkGenericWriter[optionalInt32Column](b)
+	benchmarkGenericWriter[repeatedInt32Column](b)
 }
 
 func benchmarkGenericWriter[Row generator[Row]](b *testing.B) {