Following up from #210, a significant amount of time was spent on simple loops assigning row index values in the `optionalBufferColumn` type. This PR optimizes those code paths to improve throughput of the parquet file writer.

```
name                                      old time/op  new time/op  delta
GenericWriter/optionalInt32Column/go1.18  3.30µs ± 1%  2.90µs ± 1%  -12.12%  (p=0.000 n=10+10)

name                                      old row/s    new row/s    delta
GenericWriter/optionalInt32Column/go1.18    303M ± 1%    345M ± 1%  +13.79%  (p=0.000 n=10+10)
```